# tools/swaggercodegen/swagger.json
{"swagger":"2.0","info":{"description":"A REST backend for the CSSR application","version":"0.1","title":"CSSR Backend"},"host":"localhost:8090","basePath":"/","tags":[{"name":"test-controller","description":"Test Controller"},{"name":"subscription-controller","description":"Subscription Controller"},{"name":"user-controller","description":"User Controller"},{"name":"presentation-controller","description":"Presentation Controller"}],"paths":{"/presentation":{"get":{"tags":["presentation-controller"],"summary":"getFuture","operationId":"getFutureUsingGET","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Presentation"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["presentation-controller"],"summary":"addPresentation","operationId":"addPresentationUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"pres","description":"pres","required":true,"schema":{"$ref":"#/definitions/Presentation"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Presentation"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["presentation-controller"],"summary":"modifyPresentation","operationId":"modifyPresentationUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"pres","description":"pres","required":true,"schema":{"$ref":"#/definitions/Presentation"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Presentation"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/presentation/{id}":{"get":{"tags":["presentation-controller"],"summary":"getSingle","operationId":"getSingleUsingGET","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Presentation"}},"201":{"description":"Gets data about a presentation","schema":{"type":"array","items":{"$ref":"#/definitions/Presentation"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Presentation not found in database"}}}},"/presentation/{presentationId}/subscription":{"get":{"tags":["subscription-controller"],"summary":"getSubscriptions","operationId":"getSubscriptionsUsingGET","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"presentationId","in":"path","description":"presentationId","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Subscription"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["subscription-controller"],"summary":"addSubscription","operationId":"addSubscriptionUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"presentationId","in":"path","description":"presentationId","required":true,"type":"integer","format":"int32"},{"in":"body","name":"subscription","description":"subscription","required":true,"schema":{"$ref":"#/definitions/Subscription"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Subscription"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["subscription-controller"],"summary":"modifySubscription","operationId":"modifySubscriptionUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"subscription","description":"subscription","required":true,"schema":{"$ref":"#/definitions/Subscription"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Subscription"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/presentation/{presentationId}/subscription/{subscriptionId}":{"delete":{"tags":["subscription-controller"],"summary":"deleteSingle","operationId":"deleteSingleUsingDELETE","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"presentationId","in":"path","description":"presentationId","required":true,"type":"integer","format":"int32"},{"name":"subscriptionId","in":"path","description":"subscriptionId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Subscription"}},"401":{"description":"Unauthorized"},"204":{"description":"No Content"},"403":{"description":"Forbidden"}}}},"/test":{"get":{"tags":["test-controller"],"summary":"GetTest","operationId":"GetTestUsingGET","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"name","in":"query","description":"name","required":false,"type":"string","default":"World"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/test/user":{"get":{"tags":["test-controller"],"summary":"GetUser","operationId":"GetUserUsingGET","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Principal"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/user":{"get":{"tags":["user-controller"],"summary":"getAll","operationId":"getAllUsingGET","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/User"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/user/me/resetPassword":{"post":{"tags":["user-controller"],"summary":"resetPassword","operationId":"resetPasswordUsingPOST","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}}},"definitions":{"User":{"type":"object","properties":{"displayName":{"type":"string"},"email":{"type":"string"},"passwordEnc":{"type":"string"},"tempToken":{"type":"string"},"tempTokenExpiresAt":{"$ref":"#/definitions/LocalTime"},"userId":{"type":"integer","format":"int64"}}},"Character":{"type":"object"},"LocalTime":{"type":"object","properties":{"hour":{"type":"integer","format":"int32"},"minute":{"type":"integer","format":"int32"},"nano":{"type":"integer","format":"int32"},"second":{"type":"integer","format":"int32"}}},"Presentation":{"type":"object","properties":{"abstract":{"type":"string"},"dateTime":{"type":"string","format":"date-time"},"location":{"type":"string"},"presentationId":{"type":"integer","format":"int32"},"speakerId":{"type":"integer","format":"int64"},"title":{"type":"string"}}},"Subscription":{"type":"object","properties":{"drink":{"type":"string","format":"byte"},"presentationId":{"type":"integer","format":"int32"},"sandwichType":{"$ref":"#/definitions/Character"},"subscriptionId":{"type":"integer","format":"int64"},"userId":{"type":"integer","format":"int32"}}},"Principal":{"type":"object","properties":{"name":{"type":"string"}}}}}